+++ tvdb_api.py
@@ -11,8 +11,8 @@ Example usage:
 
 >>> from tvdb_api import Tvdb
 >>> t = Tvdb()
->>> t['Lost'][4][11]['episodeName']
-u'Cabin Fever'
+>>> t['Scrubs'][1][4]['episodeName']
+'The Hotel Inspectors'
 """
 
 __author__ = "dbr/Ben"
@@ -22,39 +22,13 @@ __version__ = "3.1.0"
 import sys
 import os
 import time
-import types
 import getpass
 import tempfile
-import warnings
 import logging
-import hashlib
+from urllib.parse import quote as url_quote
 
 import requests
 import requests_cache
-
-_DEFAULT_HEADERS = requests.utils.default_headers()
-
-def _to_bytes(s, encoding='utf-8'):
-    return s if isinstance(s, bytes) else bytes(s, encoding)
-
-
-IS_PY2 = sys.version_info[0] == 2
-
-if IS_PY2:
-    user_input = raw_input  # noqa
-    from urllib import quote as url_quote  # noqa
-else:
-    from urllib.parse import quote as url_quote
-
-    user_input = input
-
-
-if IS_PY2:
-    int_types = (int, long)  # noqa
-    text_type = unicode  # noqa
-else:
-    int_types = int
-    text_type = str
 
 
 LOG = logging.getLogger("tvdb_api")
@@ -134,21 +108,6 @@ class TvdbAttributeNotFound(TvdbDataNotFound):
 
     pass
 
-
-# Backwards compatability re-bindings
-tvdb_exception = TvdbBaseException  # Deprecated, for backwards compatability
-tvdb_error = TvdbError  # Deprecated, for backwards compatability
-tvdb_userabort = TvdbUserAbort  # Deprecated, for backwards compatability
-tvdb_notauthorized = TvdbNotAuthorized  # Deprecated, for backwards compatability
-tvdb_shownotfound = TvdbShowNotFound  # Deprecated, for backwards compatability
-tvdb_seasonnotfound = TvdbSeasonNotFound  # Deprecated, for backwards compatability
-tvdb_episodenotfound = TvdbEpisodeNotFound  # Deprecated, for backwards compatability
-tvdb_resourcenotfound = TvdbResourceNotFound  # Deprecated, for backwards compatability
-tvdb_attributenotfound = TvdbAttributeNotFound  # Deprecated, for backwards compatability
-
-tvdb_invalidlanguage = TvdbError  # Unused/removed. This exists for backwards compatability.
-
-
 # UI
 
 
@@ -167,11 +126,11 @@ class BaseUI(object):
     contains the the keys "name" (human readable show name), and "sid" (the shows
     ID as on thetvdb.com). For example:
 
-    [{'name': u'Lost', 'sid': u'73739'},
-     {'name': u'Lost Universe', 'sid': u'73181'}]
+    [{'name': 'Lost', 'sid': '73739'},
+     {'name': 'Lost Universe', 'sid': '73181'}]
 
     The "selectSeries" method must return the appropriate dict, or it can raise
-    tvdb_userabort (if the selection is aborted), tvdb_shownotfound (if the show
+    TvdbUserAbort (if the selection is aborted), TvdbShowNotFound (if the show
     cannot be found).
 
     A simple example callback, which returns a random series:
@@ -187,30 +146,28 @@ class BaseUI(object):
 
     >>> from tvdb_api import Tvdb
     >>> t = Tvdb(custom_ui = RandomUI)
-    >>> random_matching_series = t['Lost']
+    >>> random_matching_series = t['Scrubs']
     >>> type(random_matching_series)
     <class 'tvdb_api.Show'>
     """
 
-    def __init__(self, config, log=None):
+    def __init__(self, config):
         self.config = config
-        if log is not None:
-            warnings.warn(
-                "the UI's log parameter is deprecated, instead use\n"
-                "use import logging; logging.getLogger('ui').info('blah')\n"
-                "The self.log attribute will be removed in the next version"
-            )
-            self.log = logging.getLogger(__name__)
 
-    def selectSeries(self, allSeries):
-        return allSeries[0]
+    def selectSeries(self, allSeries):  # noqa: N802 # Deprecated
+        """Deprecated: use `select_series` method instead
+        """
+        return self.select_series(all_series=allSeries)
+
+    def select_series(self, all_series):
+        return all_series[0]
 
 
 class ConsoleUI(BaseUI):
     """Interactively allows the user to select a show from a console based UI
     """
 
-    def _displaySeries(self, allSeries, limit=6):
+    def _display_series(self, allSeries, limit=6):
         """Helper function, lists series with corresponding ID
         """
         if limit is not None:
@@ -234,31 +191,34 @@ class ConsoleUI(BaseUI):
                 str(cshow['id']),
                 extra,
             )
-            if IS_PY2:
-                print(output.encode("UTF-8", "ignore"))
-            else:
-                print(output)
+            print(output)
 
-    def selectSeries(self, allSeries):
-        self._displaySeries(allSeries)
+    def selectSeries(self, allSeries):  # noqa: N802 # Deprecated
+        """Deprecated: use `select_series` method instead
+        """
 
-        if len(allSeries) == 1:
+        return self.select_series(all_series=allSeries)
+
+    def select_series(self, all_series):  # noqa: N802 # Deprecated
+        self._display_series(all_series)
+
+        if len(all_series) == 1:
             # Single result, return it!
             print("Automatically selecting only result")
-            return allSeries[0]
+            return all_series[0]
 
         if self.config['select_first'] is True:
             print("Automatically returning first search result")
-            return allSeries[0]
+            return all_series[0]
 
         while True:  # return breaks this loop
             try:
                 print("Enter choice (first number, return for default, 'all', ? for help):")
-                ans = user_input()
+                ans = input()
             except KeyboardInterrupt:
-                raise tvdb_userabort("User aborted (^c keyboard interupt)")
+                raise TvdbUserAbort("User aborted (^c keyboard interupt)")
             except EOFError:
-                raise tvdb_userabort("User aborted (EOF received)")
+                raise TvdbUserAbort("User aborted (EOF received)")
 
             LOG.debug('Got choice of: %s' % (ans))
             try:
@@ -267,10 +227,10 @@ class ConsoleUI(BaseUI):
                 if len(ans.strip()) == 0:
                     # Default option
                     LOG.debug('Default option, returning first series')
-                    return allSeries[0]
+                    return all_series[0]
                 if ans == "q":
                     LOG.debug('Got quit command (q)')
-                    raise tvdb_userabort("User aborted ('q' quit command)")
+                    raise TvdbUserAbort("User aborted ('q' quit command)")
                 elif ans == "?":
                     print("## Help")
                     print("# Enter the number that corresponds to the correct show.")
@@ -280,17 +240,17 @@ class ConsoleUI(BaseUI):
                     print("# q - abort tvnamer")
                     print("# Press return with no input to select first result")
                 elif ans.lower() in ["a", "all"]:
-                    self._displaySeries(allSeries, limit=None)
+                    self._display_series(all_series, limit=None)
                 else:
                     LOG.debug('Unknown keypress %s' % (ans))
             else:
                 LOG.debug('Trying to return ID: %d' % (selected_id))
                 try:
-                    return allSeries[selected_id]
+                    return all_series[selected_id]
                 except IndexError:
                     LOG.debug('Invalid show number entered!')
                     print("Invalid number (%s) selected!")
-                    self._displaySeries(allSeries)
+                    self._display_series(all_series)
 
 
 # Main API
@@ -328,7 +288,7 @@ class Show(dict):
 
     def __repr__(self):
         return "<Show %r (containing %s seasons)>" % (
-            self.data.get(u'seriesName', 'instance'),
+            self.data.get('seriesName', 'instance'),
             len(self),
         )
 
@@ -344,16 +304,16 @@ class Show(dict):
         # Data wasn't found, raise appropriate error
         if isinstance(key, int) or key.isdigit():
             # Episode number x was not found
-            raise tvdb_seasonnotfound("Could not find season %s" % (repr(key)))
+            raise TvdbSeasonNotFound("Could not find season %s" % (repr(key)))
         else:
             # If it's not numeric, it must be an attribute name, which
             # doesn't exist, so attribute error.
-            raise tvdb_attributenotfound("Cannot find attribute %s" % (repr(key)))
+            raise TvdbAttributeNotFound("Cannot find attribute %s" % (repr(key)))
 
     def aired_on(self, date):
         ret = self.search(str(date), 'firstAired')
         if len(ret) == 0:
-            raise tvdb_episodenotfound("Could not find any episodes that aired on %s" % date)
+            raise TvdbEpisodeNotFound("Could not find any episodes that aired on %s" % date)
         return ret
 
     def search(self, term=None, key=None):
@@ -381,19 +341,19 @@ class Show(dict):
         containing "my first day":
 
         >>> t['Scrubs'].search("my first day")
-        [<Episode 01x01 - u'My First Day'>]
+        [<Episode 01x01 - 'My First Day'>]
         >>>
 
         Search for "My Name Is Earl" episode named "Faked His Own Death":
 
         >>> t['My Name Is Earl'].search('Faked My Own Death', key='episodeName')
-        [<Episode 01x04 - u'Faked My Own Death'>]
+        [<Episode 01x04 - 'Faked My Own Death'>]
         >>>
 
         To search Scrubs for all episodes with "mentor" in the episode name:
 
         >>> t['scrubs'].search('mentor', key='episodeName')
-        [<Episode 01x02 - u'My Mentor'>, <Episode 03x15 - u'My Tormented Mentor'>]
+        [<Episode 01x02 - 'My Mentor'>, <Episode 03x15 - 'My Tormented Mentor'>]
         >>>
 
         # Using search results
@@ -427,7 +387,7 @@ class Season(dict):
 
     def __getitem__(self, episode_number):
         if episode_number not in self:
-            raise tvdb_episodenotfound("Could not find episode %s" % (repr(episode_number)))
+            raise TvdbEpisodeNotFound("Could not find episode %s" % (repr(episode_number)))
         else:
             return dict.__getitem__(self, episode_number)
 
@@ -437,7 +397,7 @@ class Season(dict):
 
         >>> t = Tvdb()
         >>> t['scrubs'][1].search('first day')
-        [<Episode 01x01 - u'My First Day'>]
+        [<Episode 01x01 - 'My First Day'>]
         >>>
 
         See Show.search documentation for further information on search
@@ -457,9 +417,9 @@ class Episode(dict):
         self.season = season
 
     def __repr__(self):
-        seasno = self.get(u'airedSeason', 0)
-        epno = self.get(u'airedEpisodeNumber', 0)
-        epname = self.get(u'episodeName')
+        seasno = self.get('airedSeason', 0)
+        epno = self.get('airedEpisodeNumber', 0)
+        epname = self.get('episodeName')
         if epname is not None:
             return "<Episode %02dx%02d - %r>" % (seasno, epno, epname)
         else:
@@ -469,7 +429,7 @@ class Episode(dict):
         try:
             return dict.__getitem__(self, key)
         except KeyError:
-            raise tvdb_attributenotfound("Cannot find attribute %s" % (repr(key)))
+            raise TvdbAttributeNotFound("Cannot find attribute %s" % (repr(key)))
 
     def search(self, term=None, key=None):
         """Search episode data for term, if it matches, return the Episode (self).
@@ -496,14 +456,14 @@ class Episode(dict):
         if term is None:
             raise TypeError("must supply string to search for (contents)")
 
-        term = text_type(term).lower()
+        term = str(term).lower()
         for cur_key, cur_value in self.items():
-            cur_key = text_type(cur_key)
-            cur_value = text_type(cur_value).lower()
+            cur_key = str(cur_key)
+            cur_value = str(cur_value).lower()
             if key is not None and cur_key != key:
                 # Do not search this key
                 continue
-            if cur_value.find(text_type(term)) > -1:
+            if cur_value.find(str(term)) > -1:
                 return self
 
 
@@ -528,46 +488,11 @@ class Actor(dict):
         return "<Actor %r>" % self.get("name")
 
 
-def create_key(self, request):
-    """A new cache_key algo is required as the authentication token
-    changes with each run. Also there are other header params which
-    also change with each request (e.g. timestamp). Excluding all
-    headers means that Accept-Language is excluded which means
-    different language requests will return the cached response from
-    the wrong language.
-
-    The _loadurl part checks the cache before a get is performed so
-    that an auth token can be obtained. If the response is already in
-    the cache, the auth token is not required. This prevents the need
-    to do a get which may access the host and fail because the session
-    is not yet not authorized. It is not necessary to authorize if the
-    cache is to be used thus saving host and network traffic.
-    """
-
-    if self._ignored_parameters:
-        url, body = self._remove_ignored_parameters(request)
-    else:
-        url, body = request.url, request.body
-    key = hashlib.sha256()
-    key.update(_to_bytes(request.method.upper()))
-    key.update(_to_bytes(url))
-    if request.body:
-        key.update(_to_bytes(body))
-    else:
-        if self._include_get_headers and request.headers != _DEFAULT_HEADERS:
-            for name, value in sorted(request.headers.items()):
-                # include only Accept-Language as it is important for context
-                if name in ['Accept-Language']:
-                    key.update(_to_bytes(name))
-                    key.update(_to_bytes(value))
-    return key.hexdigest()
-
-
 class Tvdb:
     """Create easy-to-use interface to name of season/episode name
     >>> t = Tvdb()
     >>> t['Scrubs'][1][24]['episodeName']
-    u'My Last Day'
+    'My Last Day'
     """
 
     def __init__(
@@ -579,11 +504,9 @@ class Tvdb:
         actors=False,
         custom_ui=None,
         language=None,
-        search_all_languages=False,
         apikey=None,
         username=None,
         userkey=None,
-        forceConnect=None,  # noqa
         dvdorder=False,
     ):
 
@@ -611,7 +534,7 @@ class Tvdb:
             via the _banners key of a Show(), for example:
 
             >>> Tvdb(banners=True)['scrubs']['_banners'].keys()
-            [u'fanart', u'poster', u'seasonwide', u'season', u'series']
+            ['fanart', 'poster', 'seasonwide', 'season', 'series']
 
         actors (True/False):
             Retrieves a list of the actors for a show. These are accessed
@@ -619,7 +542,7 @@ class Tvdb:
 
             >>> t = Tvdb(actors=True)
             >>> t['scrubs']['_actors'][0]['name']
-            u'John C. McGinley'
+            'John C. McGinley'
 
         custom_ui (tvdb_ui.BaseUI subclass):
             A callable subclass of tvdb_ui.BaseUI (overrides interactive option)
@@ -629,11 +552,6 @@ class Tvdb:
             and is also used when searching. For a complete list, call
             the `Tvdb.available_languages` method.
             Default is "en" (English).
-
-        search_all_languages (True/False):
-            By default, Tvdb will only search in the language specified using
-            the language option. When this is True, it will search for the
-            show in and language
 
         apikey (str/unicode):
             Your API key for TheTVDB. You can easily register a key with in
@@ -646,23 +564,14 @@ class Tvdb:
 
         userkey (str/unicode, or None):
             User authentication key relating to "username".
-
-        forceConnect:
-            DEPRECATED. Disabled the timeout-throttling logic. Now has no function
         """
-
-        if forceConnect is not None:
-            warnings.warn(
-                "forceConnect argument is deprecated and will be removed soon",
-                category=DeprecationWarning,
-            )
 
         self.shows = ShowContainer()  # Holds all Show classes
         self.corrections = {}  # Holds show-name to show_id mapping
 
         self.config = {}
 
-        # Ability to pull key form env-var mostly for unit-tests
+        # Ability to pull key from env-var mostly for unit-tests
         _test_key = os.getenv('TVDB_API_KEY')
         if apikey is None and _test_key is not None:
             apikey = _test_key
@@ -686,12 +595,10 @@ class Tvdb:
 
         self.config['select_first'] = select_first
 
-        self.config['search_all_languages'] = search_all_languages
-
         self.config['dvdorder'] = dvdorder
 
         if cache is True:
-            cache_dir = self._getTempDir()
+            cache_dir = self._get_temp_dir()
             LOG.debug("Caching using requests_cache to %s" % cache_dir)
             self.session = requests_cache.CachedSession(
                 expire_after=21600,  # 6 hours
@@ -699,7 +606,6 @@ class Tvdb:
                 cache_name=cache_dir,
                 include_get_headers=True,
             )
-            self.session.cache.create_key = types.MethodType(create_key, self.session.cache)
             self.session.remove_expired_responses()
             self.config['cache_enabled'] = True
         elif cache is False:
@@ -715,7 +621,6 @@ class Tvdb:
                 cache_name=os.path.join(cache, "tvdb_api"),
                 include_get_headers=True,
             )
-            self.session.cache.create_key = types.MethodType(create_key, self.session.cache)
             self.session.remove_expired_responses()
         else:
             LOG.debug("Using specified requests.Session")
@@ -763,7 +668,7 @@ class Tvdb:
             'Accept-Language': self.config['language'],
         }
 
-    def _getTempDir(self):
+    def _get_temp_dir(self):
         """Returns the [system temp dir]/tvdb_api-u501 (or
         tvdb_api-myuser)
         """
@@ -778,7 +683,7 @@ class Tvdb:
 
         return os.path.join(tempfile.gettempdir(), "tvdb_api-%s-py%s" % (uid, py_major))
 
-    def _loadUrl(self, url, data=None, recache=False, language=None):
+    def _load_url(self, url, data=None, language=None):
         """Return response from The TVDB API"""
 
         if not language:
@@ -791,7 +696,7 @@ class Tvdb:
         # encoded url is used for hashing in the cache so
         # python 2 and 3 generate the same hash
         if not self.__authorized:
-            # only authorize of we haven't before and we
+            # only authorize if we haven't before and we
             # don't have the url in the cache
             fake_session_for_key = requests.Session()
             fake_session_for_key.headers['Accept-Language'] = language
@@ -822,22 +727,21 @@ class Tvdb:
         links = r.get('links')
 
         if error:
-            if error == u'Resource not found':
-                # raise(tvdb_resourcenotfound)
+            if error == 'Resource not found':
+                # raise(TvdbResourceNotFound)
                 # handle no data at a different level so it is more specific
                 pass
-            elif error.lower() == u'not authorized':
+            elif error.lower() == 'not authorized':
                 # Note: Error string sometimes comes back as "Not authorized" or "Not Authorized"
-                raise tvdb_notauthorized()
+                raise TvdbNotAuthorized()
             elif error.startswith(u"ID: ") and error.endswith("not found"):
-                # FIXME: Refactor error out of in this method
-                raise tvdb_shownotfound("%s" % error)
+                raise TvdbShowNotFound("%s" % error)
             else:
-                raise tvdb_error("%s" % error)
+                raise TvdbError("%s" % error)
 
         if errors:
-            if errors and u'invalidLanguage' in errors:
-                # raise(tvdb_invalidlanguage(errors[u'invalidLanguage']))
+            if 'invalidLanguage' in errors:
+                # raise(TvdbError(errors['invalidLanguage']))
                 # invalidLanguage does not mean there is no data
                 # there is just less data (missing translations)
                 pass
@@ -850,7 +754,7 @@ class Tvdb:
         if links and links['next']:
             url = url.split('?')[0]
             _url = url + "?page=%s" % links['next']
-            self._loadUrl(_url, data)
+            self._load_url(_url, data=data)
 
         return data
 
@@ -862,27 +766,20 @@ class Tvdb:
         r_json = r.json()
         error = r_json.get('Error')
         if error:
-            if error == u'Not Authorized':
-                raise (tvdb_notauthorized)
+            if error == 'Not Authorized':
+                raise (TvdbNotAuthorized)
         token = r_json.get('token')
-        self.headers['Authorization'] = "Bearer %s" % text_type(token)
+        self.headers['Authorization'] = "Bearer %s" % str(token)
         self.__authorized = True
 
-    def _getetsrc(self, url, language=None):
-        """Loads a URL using caching, returns an ElementTree of the source
-        """
-        src = self._loadUrl(url, language=language)
-
-        return src
-
-    def _setItem(self, sid, seas, ep, attrib, value):
+    def _set_item(self, sid, seas, ep, attrib, value):
         """Creates a new episode, creating Show(), Season() and
         Episode()s as required. Called by _getShowData to populate show
 
         Since the nice-to-use tvdb[1][24]['name] interface
         makes it impossible to do tvdb[1][24]['name] = "name"
         and still be capable of checking if an episode exists
-        so we can raise tvdb_shownotfound, we have a slightly
+        so we can raise TvdbShowNotFound, we have a slightly
         less pretty method of setting items.. but since the API
         is supposed to be read-only, this is the best way to
         do it!
@@ -898,7 +795,7 @@ class Tvdb:
             self.shows[sid][seas][ep] = Episode(season=self.shows[sid][seas])
         self.shows[sid][seas][ep][attrib] = value
 
-    def _setShowData(self, sid, key, value):
+    def _set_show_data(self, sid, key, value):
         """Sets self.shows[sid] to a new Show instance, or sets the data
         """
         if sid not in self.shows:
@@ -911,10 +808,10 @@ class Tvdb:
         """
         series = url_quote(series.encode("utf-8"))
         LOG.debug("Searching for show %s" % series)
-        series_resp = self._getetsrc(self.config['url_getSeries'] % (series))
+        series_resp = self._load_url(self.config['url_getSeries'] % (series))
         if not series_resp:
             LOG.debug('Series result returned zero')
-            raise tvdb_shownotfound(
+            raise TvdbShowNotFound(
                 "Show-name search returned zero results (cannot find show on TVDB)"
             )
 
@@ -926,7 +823,7 @@ class Tvdb:
 
         return all_series
 
-    def _getSeries(self, series):
+    def _get_series(self, series):
         """This searches TheTVDB.com for the series name,
         If a custom_ui UI is configured, it uses this to select the correct
         series. If not, and interactive == True, ConsoleUI is used, if not
@@ -945,17 +842,17 @@ class Tvdb:
                 LOG.debug('Interactively selecting show using ConsoleUI')
                 ui = ConsoleUI(config=self.config)
 
-        return ui.selectSeries(all_series)
+        return ui.select_series(all_series)
 
     def available_languages(self):
         """Returns a list of the available language abbreviations
         which can be used in Tvdb(language="...") etc
         """
-        et = self._getetsrc("https://api.thetvdb.com/languages")
+        et = self._load_url("https://api.thetvdb.com/languages")
         languages = [x['abbreviation'] for x in et]
         return sorted(languages)
 
-    def _parseBanners(self, sid):
+    def _parse_banners(self, sid):
         """Parses banners XML, from
         http://thetvdb.com/api/[APIKEY]/series/[SERIES ID]/banners.xml
 
@@ -963,9 +860,9 @@ class Tvdb:
 
         >>> t = Tvdb(banners = True)
         >>> t['scrubs']['_banners'].keys()
-        [u'fanart', u'poster', u'seasonwide', u'season', u'series']
+        ['fanart', 'poster', 'seasonwide', 'season', 'series']
         >>> t['scrubs']['_banners']['poster']['680x1000'][35308]['_bannerpath']
-        u'http://thetvdb.com/banners/posters/76156-2.jpg'
+        'http://thetvdb.com/banners/posters/76156-2.jpg'
         >>>
 
         Any key starting with an underscore has been processed (not the raw
@@ -974,10 +871,10 @@ class Tvdb:
         This interface will be improved in future versions.
         """
         LOG.debug('Getting season banners for %s' % (sid))
-        banners_resp = self._getetsrc(self.config['url_seriesBanner'] % sid)
+        banners_resp = self._load_url(self.config['url_seriesBanner'] % sid)
         banners = {}
         for cur_banner in banners_resp.keys():
-            banners_info = self._getetsrc(self.config['url_seriesBannerInfo'] % (sid, cur_banner))
+            banners_info = self._load_url(self.config['url_seriesBannerInfo'] % (sid, cur_banner))
             for banner_info in banners_info:
                 bid = banner_info.get('id')
                 btype = banner_info.get('keyType')
@@ -1004,9 +901,9 @@ class Tvdb:
                         banners[btype][btype2][bid][new_key] = new_url
 
             banners[btype]['raw'] = banners_info
-            self._setShowData(sid, "_banners", banners)
+            self._set_show_data(sid, "_banners", banners)
 
-    def _parseActors(self, sid):
+    def _parse_actors(self, sid):
         """Parsers actors XML, from
         http://thetvdb.com/api/[APIKEY]/series/[SERIES ID]/actors.xml
 
@@ -1019,20 +916,20 @@ class Tvdb:
         >>> type(actors[0])
         <class 'tvdb_api.Actor'>
         >>> actors[0]
-        <Actor u'John C. McGinley'>
+        <Actor 'John C. McGinley'>
         >>> sorted(actors[0].keys())
-        [u'id', u'image', u'imageAdded', u'imageAuthor', u'lastUpdated', u'name', u'role',
-        u'seriesId', u'sortOrder']
+        ['id', 'image', 'imageAdded', 'imageAuthor', 'lastUpdated', 'name', 'role',
+        'seriesId', 'sortOrder']
         >>> actors[0]['name']
-        u'John C. McGinley'
+        'John C. McGinley'
         >>> actors[0]['image']
-        u'http://thetvdb.com/banners/actors/43638.jpg'
+        'http://thetvdb.com/banners/actors/43638.jpg'
 
         Any key starting with an underscore has been processed (not the raw
         data from the XML)
         """
         LOG.debug("Getting actors for %s" % (sid))
-        actors_resp = self._getetsrc(self.config['url_actorsInfo'] % (sid))
+        actors_resp = self._load_url(self.config['url_actorsInfo'] % (sid))
 
         cur_actors = Actors()
         for cur_actor_item in actors_resp:
@@ -1043,9 +940,9 @@ class Tvdb:
                         value = self.config['url_artworkPrefix'] % (value)
                 cur_actor[tag] = value
             cur_actors.append(cur_actor)
-        self._setShowData(sid, '_actors', cur_actors)
+        self._set_show_data(sid, '_actors', cur_actors)
 
-    def _getShowData(self, sid, language):
+    def _get_show_data(self, sid, language):
         """Takes a series ID, gets the epInfo URL and parses the TVDB
         XML file into the shows dict in layout:
         shows[series_id][season_number][episode_number]
@@ -1054,7 +951,7 @@ class Tvdb:
         if self.config['language'] is None:
             LOG.debug('Config language is none, using show language')
             if language is None:
-                raise tvdb_error("config['language'] was None, this should not happen")
+                raise TvdbError("config['language'] was None, this should not happen")
         else:
             LOG.debug(
                 'Configured language %s override show language of %s'
@@ -1063,30 +960,30 @@ class Tvdb:
 
         # Parse show information
         LOG.debug('Getting all series data for %s' % (sid))
-        series_info_resp = self._getetsrc(self.config['url_seriesInfo'] % sid)
+        series_info_resp = self._load_url(self.config['url_seriesInfo'] % sid)
         for tag, value in series_info_resp.items():
             if value is not None:
                 if tag in ['banner', 'fanart', 'poster']:
                     value = self.config['url_artworkPrefix'] % (value)
 
-            self._setShowData(sid, tag, value)
+            self._set_show_data(sid, tag, value)
         # set language
-        self._setShowData(sid, u'language', self.config['language'])
+        self._set_show_data(sid, 'language', self.config['language'])
 
         # Parse banners
         if self.config['banners_enabled']:
-            self._parseBanners(sid)
+            self._parse_banners(sid)
 
         # Parse actors
         if self.config['actors_enabled']:
-            self._parseActors(sid)
+            self._parse_actors(sid)
 
         # Parse episode data
         LOG.debug('Getting all episodes of %s' % (sid))
 
         url = self.config['url_epInfo'] % sid
 
-        eps_resp = self._getetsrc(url, language=language)
+        eps_resp = self._load_url(url, language=language)
 
         for cur_ep in eps_resp:
 
@@ -1121,9 +1018,9 @@ class Tvdb:
                 if value is not None:
                     if tag == 'filename':
                         value = self.config['url_artworkPrefix'] % (value)
-                self._setItem(sid, seas_no, ep_no, tag, value)
+                self._set_item(sid, seas_no, ep_no, tag, value)
 
-    def _nameToSid(self, name):
+    def _name_to_sid(self, name):
         """Takes show name, returns the correct series ID (if the show has
         already been grabbed), or grabs all episodes and returns
         the correct SID.
@@ -1133,12 +1030,12 @@ class Tvdb:
             sid = self.corrections[name]
         else:
             LOG.debug('Getting show %s' % name)
-            selected_series = self._getSeries(name)
+            selected_series = self._get_series(name)
             sid = selected_series['id']
             LOG.debug('Got %(seriesName)s, id %(id)s' % selected_series)
 
             self.corrections[name] = sid
-            self._getShowData(selected_series['id'], self.config['language'])
+            self._get_show_data(selected_series['id'], self.config['language'])
 
         return sid
 
@@ -1146,14 +1043,14 @@ class Tvdb:
         """Handles tvdb_instance['seriesName'] calls.
         The dict index should be the show id
         """
-        if isinstance(key, int_types):
+        if isinstance(key, int):
             sid = key
         else:
-            sid = self._nameToSid(key)
+            sid = self._name_to_sid(key)
             LOG.debug('Got series id %s' % sid)
 
         if sid not in self.shows:
-            self._getShowData(sid, self.config['language'])
+            self._get_show_data(sid, self.config['language'])
         return self.shows[sid]
 
     def __repr__(self):
@@ -1169,8 +1066,8 @@ def main():
     logging.basicConfig(level=logging.DEBUG)
 
     tvdb_instance = Tvdb(interactive=False, cache=False)
-    print(tvdb_instance['Lost']['seriesName'])
-    print(tvdb_instance['Lost'][1][4]['episodename'])
+    print(tvdb_instance['Scrubs']['seriesName'])
+    print(tvdb_instance['Scrubs'][1][4]['episodename'])
 
 
 if __name__ == '__main__':